Same for deployment resource model
